Refactor event handling in to separate events client
* remove featureManager instances in favor of declarative feature config
* consolidate UiAppEventsToSubscribe and UiAppEventsToTrigger types in to single UiAppEventType. Rationale is that we don't need UiAppEventsToTrigger now that we have custom events
* Features and events are no longer 1-1. Event client supports the same event handled by multiple features
* Support custom event types in subscriptions
* Change error messaging of event subscriptions: messaging happens after handshake, once per 'on' call, not when handler is executed
* Remove triggerEvent
