+++ src/constants.ts
@@ -5,20 +5,15 @@ export enum Host {
 
 export enum UiAppFeatureType {
     DASHBOARD_COG_MENU = 'dashboard_cog_menu',
-    DASHBOARD_CUSTOM_WIDGET = 'dashboard_custom_widget',
-    DASHBOARD_PAGE_CONTEXT = 'dashboard_page_context',
-    APP_ROUTING = 'app_routing'
+    DASHBOARD_CUSTOM_WIDGET = 'dashboard_custom_widget'
 }
 
-export enum UiAppEventToSubscribeType {
+export enum UiAppEventType {
+    APP_CONTEXT = 'app_context',
     DASHBOARD_COG_MENU_CONTEXT = 'dashboard_cog_menu_context',
     DASHBOARD_TIMEFRAME_CHANGE = 'dashboard_timeframe_change',
     DASHBOARD_TEMPLATE_VAR_CHANGE = 'dashboard_template_var_change',
     DASHBOARD_CUSTOM_WIDGET_OPTIONS_CHANGE = 'dashboard_custom_widget_options_change'
-}
-export enum UiAppEventToTriggerType {
-    RELOAD_FRAME = 'reload_frame',
-    OPEN_URL = 'open_url'
 }
 
 export enum IFrameApiRequestMethod {
