+++ src/index.ts
@@ -1,11 +1,11 @@
-import { DDClient } from './client';
+import { DDClient } from './client/client';
 import { Context, ClientOptions } from './types';
 
 let client: DDClient;
 
 /**
  * Initializes a client, or returns an existing one if already initialized. User can provide an optional
- * callback to be executed with app context data when it is sent from the parent.
+ * callback to b de executed with app context data when it is sent from the parent.
  */
 export const init = (
     options?: ClientOptions,
